Move stuff into src directory
