+++ lib/main_production.dart
@@ -1,4 +1,4 @@
-import 'package:betchya_frontend/app/app.dart';
+import 'package:betchya_frontend/src/app.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
