+++ lib/src/features/auth/models/email_input.dart
