+++ lib/src/features/auth/presentation/home_screen.dart
@@ -1,0 +1,45 @@
+import 'package:betchya_frontend/src/features/auth/providers/auth_provider.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+class HomeScreen extends ConsumerWidget {
+  const HomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final userAsync = ref.watch(authControllerProvider);
+
+    return userAsync.when(
+      loading: () => const Scaffold(
+        body: Center(child: CircularProgressIndicator()),
+      ),
+      error: (err, stack) => Scaffold(
+        body: Center(child: Text('Error: $err')),
+      ),
+      data: (user) => Scaffold(
+        appBar: AppBar(
+          title: const Text('Home'),
+          actions: [
+            IconButton(
+              key: const Key('home_logout_button'),
+              icon: const Icon(Icons.logout),
+              onPressed: () {
+                ref.read(authControllerProvider.notifier).signOut();
+              },
+            ),
+          ],
+        ),
+        body: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Text('Welcome, ${user?.email ?? 'User'}!'),
+              const SizedBox(height: 16),
+              const Text('You are now logged in.'),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
