--- test/app/view/app_test.dart
@@ -1,1 +1,0 @@
-void main() {}
