--- test/features/auth/controllers/login_form_controller_test.dart
@@ -1,154 +1,0 @@
-import 'package:betchya_frontend/features/auth/controllers/login_form_controller.dart';
-import 'package:betchya_frontend/features/auth/models/email_input.dart';
-import 'package:betchya_frontend/features/auth/models/password_input.dart';
-import 'package:betchya_frontend/features/auth/providers/auth_provider.dart';
-import 'package:betchya_frontend/features/auth/repository/auth_repository.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:formz/formz.dart';
-import 'package:mocktail/mocktail.dart';
-
-class MockAuthController extends Mock implements AuthController {}
-
-class MockAuthRepository extends Mock implements AuthRepository {}
-
-void main() {
-  late MockAuthController mockAuthController;
-  late LoginFormController controller;
-
-  const testEmail = 'test@example.com';
-  const testPassword = 'Test1234';
-
-  setUp(() {
-    mockAuthController = MockAuthController();
-    controller = LoginFormController(mockAuthController);
-  });
-
-  group('LoginFormController', () {
-    test('initial state is correct', () {
-      expect(controller.state.email, const EmailInput.pure());
-      expect(controller.state.password, const PasswordInput.pure());
-      expect(controller.state.status, FormzStatus.pure);
-      expect(controller.state.error, isNull);
-      expect(controller.state.isSubmitting, isFalse);
-    });
-
-    test('loginFormControllerProvider provides a LoginFormController', () {
-      final mockAuthRepository = MockAuthRepository();
-      final container = ProviderContainer(
-        overrides: [
-          authRepositoryProvider.overrideWithValue(mockAuthRepository),
-        ],
-      );
-      final controller = container.read(loginFormControllerProvider.notifier);
-      expect(controller, isA<LoginFormController>());
-    });
-
-    group('emailChanged', () {
-      test('updates email and validates form', () {
-        // Act
-        controller.emailChanged(testEmail);
-
-        // Assert
-        expect(controller.state.email.value, testEmail);
-        expect(controller.state.status, FormzStatus.invalid);
-      });
-    });
-
-    group('passwordChanged', () {
-      test('updates password and validates form', () {
-        // Act
-        controller.passwordChanged(testPassword);
-
-        // Assert
-        expect(controller.state.password.value, testPassword);
-        expect(controller.state.status, FormzStatus.invalid);
-      });
-    });
-
-    group('submit', () {
-      test('does nothing when form is invalid', () async {
-        // Act
-        await controller.submit();
-
-        // Assert
-        verifyNever(
-          () => mockAuthController.signIn(
-            email: any(named: 'email'),
-            password: any(named: 'password'),
-          ),
-        );
-      });
-
-      test('calls signIn when form is valid', () async {
-        // Arrange
-        // Mock the signIn method first
-        when(
-          () => mockAuthController.signIn(
-            email: testEmail,
-            password: testPassword,
-          ),
-        ).thenAnswer((_) async => {});
-
-        // Set up valid form state after mocking
-        controller
-          ..emailChanged(testEmail)
-          ..passwordChanged(testPassword)
-
-          // Force form to be valid by directly setting the status
-          // This is needed because our form validation might be too strict
-          // for testing
-          ..state = controller.state.copyWith(
-            status: FormzStatus.valid,
-          );
-
-        // Act
-        await controller.submit();
-
-        // Assert
-        verify(
-          () => mockAuthController.signIn(
-            email: testEmail,
-            password: testPassword,
-          ),
-        ).called(1);
-        expect(controller.state.isSubmitting, isFalse);
-      });
-
-      test('handles signIn error', () async {
-        // Arrange
-        final exception = Exception('Sign in failed');
-
-        // Mock the signIn method to throw an exception
-        when(
-          () => mockAuthController.signIn(
-            email: testEmail,
-            password: testPassword,
-          ),
-        ).thenThrow(exception);
-
-        // Set up form state after mocking
-        controller
-          ..emailChanged(testEmail)
-          ..passwordChanged(testPassword)
-
-          // Force form to be valid by directly setting the status
-          // This is needed because our form validation might be too strict
-          // for testing
-          ..state = controller.state.copyWith(
-            status: FormzStatus.valid,
-          );
-
-        // Act & Assert
-        expect(
-          () => controller.submit(),
-          throwsA(exception),
-        );
-
-        // Verify the state after error
-        expect(controller.state.error, exception.toString());
-        expect(controller.state.isSubmitting, isFalse);
-      });
-    });
-  });
-}
