--- test/features/auth/controllers/sign_up_form_controller_test.dart
@@ -1,224 +1,0 @@
-import 'package:betchya_frontend/features/auth/controllers/sign_up_form_controller.dart';
-import 'package:betchya_frontend/features/auth/models/confirm_password_input.dart';
-import 'package:betchya_frontend/features/auth/models/email_input.dart';
-import 'package:betchya_frontend/features/auth/models/full_name_input.dart';
-import 'package:betchya_frontend/features/auth/models/password_input.dart';
-import 'package:betchya_frontend/features/auth/repository/auth_repository.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:formz/formz.dart';
-import 'package:mocktail/mocktail.dart';
-
-class MockAuthRepository extends Mock implements AuthRepository {}
-
-void main() {
-  late MockAuthRepository mockAuthRepository;
-  late SignUpFormController controller;
-
-  const testEmail = 'test@example.com';
-  const testPassword = 'Test1234';
-
-  setUp(() {
-    mockAuthRepository = MockAuthRepository();
-    controller = SignUpFormController(mockAuthRepository);
-  });
-
-  group('SignUpFormController', () {
-    test('initial state is correct', () {
-      expect(controller.state.email, const EmailInput.pure());
-      expect(controller.state.password, const PasswordInput.pure());
-      expect(controller.state.fullName, const FullNameInput.pure());
-      expect(
-        controller.state.confirmPassword,
-        const ConfirmPasswordInput.pure(),
-      );
-      expect(controller.state.status, FormzStatus.pure);
-      expect(controller.state.error, isNull);
-      expect(controller.state.isSubmitting, isFalse);
-    });
-
-    group('emailChanged', () {
-      test('updates email and validates form', () {
-        // Act
-        controller.emailChanged(testEmail);
-
-        // Assert
-        expect(controller.state.email.value, testEmail);
-        expect(controller.state.status, FormzStatus.invalid);
-      });
-    });
-
-    group('passwordChanged', () {
-      test('updates password and validates form', () {
-        controller.passwordChanged(testPassword);
-        expect(controller.state.password.value, testPassword);
-        expect(controller.state.status, FormzStatus.invalid);
-      });
-
-      test('changing password re-validates confirmPassword', () {
-        // Set confirmPassword to match old password
-        controller
-          ..passwordChanged('OldPassword1!')
-          ..confirmPasswordChanged('OldPassword1!');
-        expect(controller.state.confirmPassword.error, isNull);
-        // Change password, confirmPassword is now out of sync
-        controller.passwordChanged('NewPassword2!');
-        expect(controller.state.confirmPassword.error, isNotNull);
-        // Now update confirmPassword to match
-        controller.confirmPasswordChanged('NewPassword2!');
-        expect(controller.state.confirmPassword.error, isNull);
-      });
-    });
-
-    group('confirmPasswordChanged', () {
-      test('sets error if confirm password does not match', () {
-        controller
-          ..passwordChanged('Password1!')
-          ..confirmPasswordChanged('WrongPassword');
-        expect(controller.state.confirmPassword.error, isNotNull);
-      });
-      test('no error if confirm password matches', () {
-        controller
-          ..passwordChanged('Password1!')
-          ..confirmPasswordChanged('Password1!');
-        expect(controller.state.confirmPassword.error, isNull);
-      });
-    });
-
-    test(
-        'form is valid only if all fields including confirm password are '
-        'valid and match', () {
-      controller
-        ..emailChanged('valid@email.com')
-        ..passwordChanged('Valid123!')
-        ..confirmPasswordChanged('Valid123!')
-        ..fullNameChanged('John Doe');
-      expect(controller.state.status, FormzStatus.valid);
-      // Now break confirm password
-      controller.confirmPasswordChanged('WrongPassword');
-      expect(controller.state.status, isNot(FormzStatus.valid));
-    });
-
-    group('fullNameChanged', () {
-      test('updates fullName and validates form', () {
-        // Act
-        controller.fullNameChanged('John Doe');
-
-        // Assert
-        expect(controller.state.fullName.value, 'John Doe');
-        expect(controller.state.status, FormzStatus.invalid);
-      });
-    });
-
-    group('submit', () {
-      test('does nothing when form is invalid', () async {
-        // Act
-        await controller.submit();
-
-        // Assert
-        verifyNever(
-          () => mockAuthRepository.signUp(
-            email: any(named: 'email'),
-            password: any(named: 'password'),
-          ),
-        );
-      });
-
-      test('calls signUp when form is valid', () async {
-        // Arrange
-        // Mock the signUp method first
-        when(
-          () => mockAuthRepository.signUp(
-            email: testEmail,
-            password: testPassword,
-          ),
-        ).thenAnswer((_) async => null);
-
-        // Set up form state after mocking
-        controller
-          ..emailChanged(testEmail)
-          ..passwordChanged(testPassword)
-          ..fullNameChanged('John Doe')
-
-          // Force form to be valid by directly setting the status
-          // This is needed because our form validation might be too strict
-          // for testing
-          ..state = controller.state.copyWith(
-            status: FormzStatus.valid,
-          );
-
-        // Act
-        await controller.submit();
-
-        // Assert
-        verify(
-          () => mockAuthRepository.signUp(
-            email: testEmail,
-            password: testPassword,
-          ),
-        ).called(1);
-        expect(controller.state.isSubmitting, isFalse);
-      });
-
-      test('handles signUp error', () async {
-        final exception = Exception('Sign up failed');
-        when(
-          () => mockAuthRepository.signUp(
-            email: testEmail,
-            password: testPassword,
-          ),
-        ).thenThrow(exception);
-
-        controller
-          ..emailChanged(testEmail)
-          ..passwordChanged(testPassword)
-          ..fullNameChanged('John Doe')
-          ..state = controller.state.copyWith(status: FormzStatus.valid);
-
-        await controller.submit();
-
-        expect(controller.state.error, exception.toString());
-        expect(controller.state.isSubmitting, isFalse);
-      });
-      group('consentChanged', () {
-        test('updates the consent value', () {
-          // consent starts as true
-          expect(controller.state.consent, isTrue);
-
-          // Change to false
-          controller.consentChanged(value: false);
-          expect(controller.state.consent, isFalse);
-
-          // Change back to true
-          controller.consentChanged(value: true);
-          expect(controller.state.consent, isTrue);
-        });
-      });
-
-      group('validateAll', () {
-        test('sets all fields to dirty and updates status', () {
-          // fullName is dirty
-          controller..fullNameChanged('John Doe')
-          // email is dirty
-          ..emailChanged(testEmail)
-          // password is dirty
-          ..passwordChanged(testPassword)
-          // confirmPassword is dirty
-          ..confirmPasswordChanged(testPassword)
-          // consent is dirty
-          ..consentChanged(value: false)
-
-          // Validate all
-          ..validateAll();
-
-          // Assert all fields are dirty and status is updated
-          expect(controller.state.fullName.pure, isFalse);
-          expect(controller.state.email.pure, isFalse);
-          expect(controller.state.password.pure, isFalse);
-          expect(controller.state.confirmPassword.pure, isFalse);
-          expect(controller.state.consent, isFalse);
-          expect(controller.state.status, FormzStatus.invalid);
-        });
-      });
-    });
-  });
-}
