--- test/features/auth/presentation/home_screen_test.dart
@@ -1,48 +1,0 @@
-import 'package:betchya_frontend/features/auth/presentation/home_screen.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  setUp(() {
-    // Initialize any global mocks or dependencies here if needed
-  });
-
-  tearDown(() {
-    // Clean up after tests if needed
-  });
-
-  Future<void> pumpHomeScreen(
-    WidgetTester tester, {
-    List<Override> overrides = const [],
-  }) async {
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: overrides,
-        child: const HomeScreen(),
-      ),
-    );
-    await tester.pumpAndSettle();
-  }
-
-  group('HomeScreen', () {
-    testWidgets('renders main UI elements', (tester) async {
-      // Should verify all main home screen elements are present
-    });
-
-    testWidgets('displays user info correctly', (tester) async {
-      // Should display correct user information
-    });
-
-    testWidgets('navigates to other screens via buttons', (tester) async {
-      // Should navigate to other screens when buttons are tapped
-    });
-
-    testWidgets('shows loading indicator when fetching data', (tester) async {
-      // Should show a loading spinner when fetching data
-    });
-
-    testWidgets('shows error message on data fetch failure', (tester) async {
-      // Should display an error message if fetching data fails
-    });
-  });
-}
