--- test/features/auth/presentation/login_screen_test.dart
@@ -1,52 +1,0 @@
-import 'package:betchya_frontend/features/auth/presentation/login_screen.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  setUp(() {
-    // Initialize any global mocks or dependencies here if needed
-  });
-
-  tearDown(() {
-    // Clean up after tests if needed
-  });
-
-  Future<void> pumpLoginScreen(
-    WidgetTester tester, {
-    List<Override> overrides = const [],
-  }) async {
-    await tester.pumpWidget(
-      ProviderScope(
-        overrides: overrides,
-        child: const LoginScreen(),
-      ),
-    );
-    await tester.pumpAndSettle();
-  }
-
-  group('LoginScreen', () {
-    testWidgets('renders all required input fields', (tester) async {
-      // Should verify all login fields are present
-    });
-
-    testWidgets('shows validation errors for invalid input', (tester) async {
-      // Should show error messages for invalid data
-    });
-
-    testWidgets('enables login button only when form is valid', (tester) async {
-      // Should only enable the login button when all fields are valid
-    });
-
-    testWidgets('shows loading indicator when logging in', (tester) async {
-      // Should show a loading spinner when login is in progress
-    });
-
-    testWidgets('navigates to home on successful login', (tester) async {
-      // Should navigate to home screen after successful login
-    });
-
-    testWidgets('shows error on failed login', (tester) async {
-      // Should display an error message if login fails
-    });
-  });
-}
