+++ test/helpers/pump_app.dart
@@ -1,4 +1,4 @@
-import 'package:betchya_frontend/l10n/l10n.dart';
+import 'package:betchya_frontend/src/l10n/l10n.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
